backend/requirements.txt
# FURG Backend Requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25

# Authentication & Security
python-jose[cryptography]==3.3.0
cryptography==42.0.0

# HTTP Client
httpx==0.26.0

# AI/ML
anthropic==0.18.0

# Plaid Integration
plaid-python==18.0.0

# Data Processing
python-dateutil==2.8.2

# Environment
python-dotenv==1.0.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
